Simplify reference processing
